Move to e2e-full
