+++ site/gatsby-site/playwright/config.ts
@@ -11,6 +11,7 @@ const config: ConfigType = {
     E2E_ADMIN_USERNAME: process.env.E2E_ADMIN_USERNAME!,
     IS_EMPTY_ENVIRONMENT: process.env.IS_EMPTY_ENVIRONMENT ?? '',
     AVAILABLE_LANGUAGES: process.env.GATSBY_AVAILABLE_LANGUAGES ?? '',
+    MONGODB_TRANSLATIONS_CONNECTION_STRING: process.env.MONGODB_TRANSLATIONS_CONNECTION_STRING!,
 }
 
 Object.keys(config).forEach((key) => {
